Add JavaCPP
Unfortunately, the javadoc.io server does not support our
Proxy/ProxyReverse scheme. When configured as such, we see:

  curl -I https://javadoc.scijava.org/JavaCPP/index.html
  HTTP/1.1 303 See Other
  Date: Thu, 14 Nov 2019 13:42:26 GMT
  Server: cloudflare
  Location: /doc/org.bytedeco/javacpp/1.4.2/index.html
  Referrer-Policy: origin-when-cross-origin, strict-origin-when-cross-origin

There may be a way to make it work, but I am not an Apache expert,
and javadoc.io seems likely to be a stable resource as time goes on,
so for now let's just link to it directly.

One good thing is that unlike javadoc.scijava.org, javadoc.io has
the advantage of being able to toggle between component versions.